requirements.txt
asgiref==3.6.0
Django==4.2
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
furl==2.1.3
orderedmultidict==1.0.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
